added resource sections
