+++ apps/web/src/components/GoToCommunity/resourceSections.ts
@@ -1,3 +1,5 @@
+import { ResourceSectionType } from 'apps/web/src/components/GoToCommunity/ResourceSection';
+
 export const resourceSections: ResourceSectionType[] = [
   {
     title: 'Resources to BUILD your project',
@@ -6,18 +8,22 @@ export const resourceSections: ResourceSectionType[] = [
       {
         title: 'Smart Wallet',
         description: 'Help users create an account in seconds without apps or extensions',
+        href: '/about',
       },
       {
-        title: 'Smart Wallet',
+        title: 'Resource #2',
         description: 'Help users create an account in seconds without apps or extensions',
+        href: '/about',
       },
       {
-        title: 'Smart Wallet',
+        title: '#3',
         description: 'Help users create an account in seconds without apps or extensions',
+        href: '/about',
       },
       {
-        title: 'Smart Wallet',
+        title: 'Fourth Resource',
         description: 'Help users create an account in seconds without apps or extensions',
+        href: '/about',
       },
     ],
   },
@@ -28,30 +34,49 @@ export const resourceSections: ResourceSectionType[] = [
       {
         title: 'Smart Wallet',
         description: 'Help users create an account in seconds without apps or extensions',
+        href: '/about',
       },
       {
         title: 'Something Else',
         description: 'Help users create an account in seconds without apps or extensions',
+        href: '/about',
       },
       {
         title: 'A third thing',
         description: 'Help users create an account in seconds without apps or extensions',
+        href: '/about',
       },
       {
         title: 'Thing #4',
         description: 'Help users create an account in seconds without apps or extensions',
+        href: '/about',
+      },
+    ],
+  },
+  {
+    title: 'Resources to FUND your project',
+    description: 'Get money money money, get money money!',
+    cards: [
+      {
+        title: 'Smart Wallet',
+        description: 'Help users create an account in seconds without apps or extensions',
+        href: '/about',
+      },
+      {
+        title: 'Something Else',
+        description: 'Help users create an account in seconds without apps or extensions',
+        href: '/about',
+      },
+      {
+        title: 'A third thing',
+        description: 'Help users create an account in seconds without apps or extensions',
+        href: '/about',
+      },
+      {
+        title: 'Thing #4',
+        description: 'Help users create an account in seconds without apps or extensions',
+        href: '/about',
       },
     ],
   },
 ];
-
-export type ResourceCard = {
-  title: string;
-  description: string;
-};
-
-export type ResourceSectionType = {
-  title: string;
-  description: string;
-  cards: ResourceCard[];
-};
